+++ max_div/_cli/bm_solver/_benchmark_solver_run.py
@@ -1,4 +1,4 @@
-from .run_initialization import benchmark_initialization_strategies
+from .benchmark_initialization import benchmark_initialization_strategies
 
 
 def run_solver_benchmark(name: str, markdown: bool):